Migrate back haproxy validation to tripleo-validations
This patch moves back the haproxy validation from validations-common to
tripleo-validations. It has been renamed in tripleo-haproxy and the role
is called now tripleo_haproxy.

This patch also fix the host targeting in order to only execute this
validation on nodes where haproxy is really installed and configured.

Note that the haproxy validations hosted in validations-common will be
removed in a follow up patch.

Closes-Bug: #1926024

Co-Authored-By: Michele Baldessari <michele@redhat.com>

Signed-off-by: Gael Chamoulaud (Strider) <gchamoul@redhat.com>
Change-Id: I9869afd60342fdb0aca69a313c1c6e35e0947fb8