Add groups.yaml to the openstack-tox-jobs files
This patch adds the `groups.yaml` to the list of files which triggers
the documentation build.

This patch also adds some words in the documentation in order to notify
the user that he/she/it will have to add a new entry in the doc
index.rst file when adding a new validations group.

Signed-off-by: Gael Chamoulaud (Strider) <gchamoul@redhat.com>
Change-Id: Ib4a92f98821bcecd4e2f84ee0b6d3445152ee8ce