Introduce new Ansible validation_init role.
This patch migrates the new role addition playbook into a proper Ansible
role with proper molecule tests. This new role will be responsible for
creating a new validation.

It is good to mention that the process won't change and it will be
fully transparent for the user.

To create a new role in tripleo-validations:

$ cd tripleo-validations/
$ export ANSIBLE_ROLES_PATH="${PWD}/roles"
$ ansible-playbook -i localhost, role-addition.yml -e validation_init_role_name=${NEWROLENAME}

Signed-off-by: Gael Chamoulaud (Strider) <gchamoul@redhat.com>
Change-Id: I219f4054e6b854306719b9c91a0738c526f07666
(cherry picked from commit 83ba5e32fdf617d2669b5ed83517729b470a7cd5)