Introducing the oslo_config_validator role
This role is intended to leverage the `oslo-config-validator` on each one
of the configuration files found on a deployment. The goal is to quickly
catch erroneous configurations.

When called manually, it will also be possible to generate a report
returning all the differences between the current configuration and the
default configuration

One last thing, it will also validate the values of some specific
settings, based on the invalid_settings dict in the configuration.
Because of that, we moved some checks in nova_event_callback to the
oslo_config_validator role.

NOTE(dvd): Removing backup-and-restore group from this branch as it's
not available prior to Wallaby.

Depends-On: https://review.opendev.org/c/openstack/oslo.config/+/794815
Related: https://bugzilla.redhat.com/show_bug.cgi?id=1940815
Resolves: rhbz#1940815
Change-Id: Id047fe378cf512b985c9d7478a991f11b280102b
(cherry picked from commit 4aaca32f19af6426a6f1eb6ea896e8d10458928d)