Moving setting validation from nova_event_callback
The newly introduced role oslo_config_validator is probably a
better framework to validate settings as it's centralized and
will most probably be called by sosreport for troubleshooting
purpose. It only makes sense to move some setting validations
to this new role when possible.

This patch also adds the nova_event_callback role documentation in
sphinx.

Closes-Bug: #1934609

Change-Id: Iac8126429773d3b63d17ad851d24e8c6f498f557