[validation_init] Remove the creation of a README.md file for a new validation
This patch removes the creation of a role README.md file during the
creation of a new validation with the validation_init role and adds
documentation directly in the sphinx role documentation file. All
variables are documented using comment(s) and will be rendered
automatically by the ansible-autodoc.py script during the sphinx build.

Signed-off-by: Gael Chamoulaud (Strider) <gchamoul@redhat.com>
Change-Id: I4341483264dbe657066f03a132ee83943e4aa823