Add zuul.d/molecule.yaml file back in the yamllint ignore list
The validation_init role auto generates the molecule.yaml file when
adding a new role. But, due to [1][2], the to_nice_yaml filter is not
correctly indenting lists and will break the indentation every time a
new validation will be created.

This patch adds this file to the yamllint ignore list and provides a new
molecule.yaml sorted by job names as the validation_init role will
generate in a follow up patch.

Moreover, this patch enables the oslo_config_validator molecule job in
the check/gate as it was not added in the initial patch[3].

[1] https://github.com/ansible/ansible/issues/48865
[2] https://github.com/yaml/pyyaml/issues/234
[3] https://review.opendev.org/c/openstack/tripleo-validations/+/789698

Signed-off-by: Gael Chamoulaud (Strider) <gchamoul@redhat.com>
Change-Id: Iac63dc6f161a0e885110a246b68001522a6f62e3