[validation_init] Add ability to create new validation w/o molecule test
This patch adds a new variable called `validation_init_molecule` to the
`validation_init` role in order to create a new validation with or
without Molecule test.

If sets to `false` (by default `true`), the molecule directory for the
new role will be removed and no CI Job will be added into the
zuul.d/molecule.yaml file.

Signed-off-by: Gael Chamoulaud (Strider) <gchamoul@redhat.com>
Change-Id: Ia068957eed5e868b40c2386e3844d7a51bf61bb6