Fix Inventory Generation without Overcloud deployed
The utils.list_plan_and_stack function returns None in case of exceptions
or and empty list if no stacks or no plans have been found in the
environment.

This patch fixes the conditional to allow the tripleo-ansible-inventory
script to generate only the undercloud Inventory when no Overcloud has
been deployed.

Related-bz: rhbz#1987318

Signed-off-by: Gael Chamoulaud (Strider) <gchamoul@redhat.com>
Change-Id: If0dc3300fbd8cf245c0477be782110d897344860