feat(removal): tox-ansible support
This patch removes the support of tox-ansible from
tripleo-validations. Due to several issues with pip dependencies
management and our requirements files, the installation of this latter
doesn't work.

This patch also removes the `tripleo-linters` anchor and nodeset line
for the `openstack-tox-linters` check job as they are not used or useful
here.

JIRA: #VALFRWK702

Signed-off-by: Gael Chamoulaud (Strider) <gchamoul@redhat.com>
Change-Id: I02d70e9eaf112d9d9137a020c8b236c150a12149