Remove references to novajoin-server and novajoin-notifier
As of Wallaby, novajoin-server and novajoin-notifier are no longer
used.  Instead, enrollment is done through the tripleo-ipa ansible
role.

So, we can remove any checks on these services (which ran in
containers).  The only vetiges of these services is the location
of the novajoin keytab, which is a file on the undercloud.

Change-Id: Id2f0c7d961c6249193f0eea4c02e3063aa49fbd7
(cherry picked from commit 52f3fbd75a664e01249f65b0ecbf01fd233d21b1)