Instruct container-status validation to accept 137, 142 and 143 exit code status
This patch makes the validations aware of those exit code status and
accept them as not a validation failure.

Below, the upstream patches for tripleo-ansible[1] and paunch[2]:

[1] - https://review.opendev.org/q/I8f19a80016a67ccad0371c5d108516aec640f031
[2] - https://review.opendev.org/q/Iffcfc8bd18a999ae6921a4131d40241df40050f1

PS1: This patch needs to be back ported to Wallaby, but also to Victoria
and Ussuri for keeping the history clean.

PS2: This validation depends on the tripleo-ansible-inventory[3] script
to get the overcloud `container_cli` var, but that latter is not operable
anymore on the stable branches above. In conclusion, this validation
will be considered as broken until the inventory is fixed.

Partial-fix: rhbz#2040667

Signed-off-by: Gael Chamoulaud (Strider) <gchamoul@redhat.com>
Change-Id: I6febdcae790075a8b3ac1323e8b1c189882005b3
(cherry picked from commit 8989c8c03e789213ed09180abc039a657ee488ce)
(cherry picked from commit 1b8766d14ee420682332218834e044885aa50d8b)