Redefining the irrelevant files list to ensure execution graph stability
The default irrelevant file definition was leading
to potential edge case when jobs were requested, with their dependencies
being skipped at the same time.
This caused discontinuity of the execution graph and immediety failure of
the check pipeline.

This patch resolves the issue by redefining the irrelevant files list
for selected jobs.

Signed-off-by: Jiri Podivin <jpodivin@redhat.com>
Change-Id: I3585ad40741db85845fab86987045c1ac10eb084