Substituting 'localhost' for wildcard '*' in image-serve validation
Recent changes to the httpd config template[0] have led to
failure of the image-serve validation during the "Ensure registry does answer"
task as the '*' host was not reachable.

Substitution of the 'localhost' for '*' should resolve the issue.

Same modification was applied to the "Ensure port is open" task
for the sake of consistency.

Closes-Bug: #1959864

[0] https://opendev.org/openstack/tripleo-ansible/commit/9b2ffd10dccefcb5462a12ffb86d1eaf8b64adb5
Signed-off-by: Jiri Podivin <jpodivin@redhat.com>
Change-Id: I90171b424f0678001c2ca53a1fee819993fff6f5