Check the time synchronization and list the NTP servers chrony is trying to access
Time synchronization is essential for successful deployment. The chrony daemon is used for connection with the NTP pool for receiving time data, if the connection is unsuccessful the deployment fails. This validation is recommended to run if the deployment is failing on time synchronization task. This validation checks if the time is synchronized and prints NTP servers chrony is trying to access.

Signed-off-by: Veronika Fisarova <vfisarov@redhat.com>
Change-Id: Ia25673018b9ae1a44a068291de182ebf827466a1