Triggering overcloud container CLI discovery task on allovercloud group
The overcloud container CLI was previously only discovered if overcloud group
was present in the inventory. As one of the validations targets,
the 'allovercloud' group over took much of the previous 'overcloud' group responsibilities.

This could lead to an edge case when no overcloud container CLI was discovered on overcloud nodes,
and the undercloud cli was never sought. The relevant variable remained
unfilled and caused eventual syntax error during evaluation of a conditional
in the following task.

Resolves: rhbz#2127928

Signed-off-by: Jiri Podivin <jpodivin@redhat.com>
Change-Id: I84a7c20a41698d57a379720ef0424f5220bb9edd
(cherry picked from commit 29ad8ab3421776f906f98fac3f3efcc680a11e11)